# docs/swagger.yaml
definitions:
  APIAuthorization:
    description: APIAuthorization defines mock auth parameters
    properties:
      format:
        type: string
        x-go-name: Format
      in:
        type: string
        x-go-name: In
      name:
        type: string
        x-go-name: Name
      scheme:
        type: string
        x-go-name: Scheme
      type:
        type: string
        x-go-name: Type
      url:
        type: string
        x-go-name: URL
    type: object
    x-go-package: github.com/bhatti/api-mock-service/internal/types
...
  APIRequest:
    description: APIRequest defines mock request for APIs
    properties:
      assert_contents_pattern:
        description: AssertContentsPattern for request optionally
        type: string
        x-go-name: AssertContentsPattern
      assert_headers_pattern:
        additionalProperties:
          type: string
        description: AssertHeadersPattern for mock response
        type: object
        x-go-name: AssertHeadersPattern
      assert_query_params_pattern:
        additionalProperties:
          type: string
        description: AssertQueryParamsPattern for the API
        type: object
        x-go-name: AssertQueryParamsPattern
      assertions:
        description: Assertions for validating response
        items:
          type: string
        type: array
        x-go-name: Assertions
      contents:
        description: Contents for request optionally
        type: string
        x-go-name: Contents
      example_contents:
        description: ExampleContents sample for request optionally
        type: string
        x-go-name: ExampleContents
      headers:
        additionalProperties:
          type: string
        description: Headers for mock response
        type: object
        x-go-name: Headers
      path_params:
        additionalProperties:
          type: string
        description: PathParams sample for the API
        type: object
        x-go-name: PathParams
      query_params:
        additionalProperties:
          type: string
        description: QueryParams sample for the API
        type: object
        x-go-name: QueryParams
    type: object
    x-go-package: github.com/bhatti/api-mock-service/internal/types
  APIResponse:
    description: APIResponse defines mock response for APIs
    properties:
      assert_contents_pattern:
        description: AssertContentsPattern for request optionally
        type: string
        x-go-name: AssertContentsPattern
      assert_headers_pattern:
        additionalProperties:
          type: string
        description: AssertHeadersPattern for mock response
        type: object
        x-go-name: AssertHeadersPattern
      assertions:
        description: Assertions for validating response
        items:
          type: string
        type: array
        x-go-name: Assertions
      contents:
        description: Contents for request
        type: string
        x-go-name: Contents
      contents_file:
        description: ContentsFile for request
        type: string
        x-go-name: ContentsFile
      example_contents:
        description: ExampleContents sample for response optionally
        type: string
        x-go-name: ExampleContents
      headers:
        additionalProperties:
          items:
            type: string
          type: array
        description: Headers for mock response
        type: object
        x-go-name: Headers
      pipe_properties:
        description: PipeProperties to extract properties from response
        items:
          type: string
        type: array
        x-go-name: PipeProperties
      status_code:
        description: StatusCode for response
        format: int64
        type: integer
        x-go-name: StatusCode
    type: object
    x-go-package: github.com/bhatti/api-mock-service/internal/types
  APIScenario:
    description: APIScenario defines mock scenario for APIs
    properties:
      authentication:
        additionalProperties:
          $ref: '#/definitions/APIAuthorization'
        description: Authentication for the API
        type: object
        x-go-name: Authentication
      base_url:
        description: BaseURL of remote server
        type: string
        x-go-name: BaseURL
      description:
        description: Description of scenario
        type: string
        x-go-name: Description
      group:
        description: Group of scenario
        type: string
        x-go-name: Group
      method:
        $ref: '#/definitions/MethodType'
      name:
        description: Name to uniquely identify the scenario
        type: string
        x-go-name: Name
      order:
        description: Order of scenario
        format: int64
        type: integer
        x-go-name: Order
      path:
        description: Path for the API (excluding query params)
        type: string
        x-go-name: Path
      predicate:
        description: Predicate for the request
        type: string
        x-go-name: Predicate
      request:
        $ref: '#/definitions/APIRequest'
      response:
        $ref: '#/definitions/APIResponse'
      tags:
        description: Tags of scenario
        items:
          type: string
        type: array
        x-go-name: Tags
      wait_before_reply:
        $ref: '#/definitions/Duration'
    type: object
    x-go-package: github.com/bhatti/api-mock-service/internal/types
  Duration:
    description: |-
      A Duration represents the elapsed time between two instants
      as an int64 nanosecond count. The representation limits the
      largest representable duration to approximately 290 years.
    format: int64
    type: integer
    x-go-package: time
  MethodType:
    description: MethodType for HTTP
    type: string
    x-go-package: github.com/bhatti/api-mock-service/internal/types
...
paths:
...
  /_oapi:
    post:
      description: Creates new api scenarios based on Open API v3
      operationId: postMockOAPIScenario
      parameters:
      - in: body
        name: Body
        schema:
          items:
            format: uint8
            type: integer
          type: array
      responses:
        "200":
          $ref: '#/responses/apiScenarioOAPIResponse'
      tags:
      - open-api
...
responses:
...
  apiScenarioOAPIResponse:
    description: APIScenario body for update
    schema:
      $ref: '#/definitions/APIScenario'
...
swagger: "2.0"